Make ShaderStage an enum class

+++ filament/src/MaterialParser.cpp
@@ -309,7 +309,7 @@ bool MaterialParser::getReflectionMode(ReflectionMode* value) const noexcept {
 bool MaterialParser::getShader(ShaderContent& shader,
         ShaderModel shaderModel, Variant variant, ShaderType stage) noexcept {
     return mImpl.mMaterialChunk.getShader(shader,
-            mImpl.mBlobDictionary, (uint8_t)shaderModel, variant, stage);
+            mImpl.mBlobDictionary, uint8_t(shaderModel), variant, uint8_t(stage));
 }
 
 // ------------------------------------------------------------------------------------------------
